Revert "moving config file to temp location before release (#110)"
This reverts commit 1ff525bfc13ce8379445bb6a0a658b833fedf767.
